Add v2 method
